refactor: replace os.path with pathlib for improved path handling

+++ preprocessing/file_structures/make_stack_from_single_tifs.py
@@ -2,6 +2,7 @@ from tifffile import *
 import os, tqdm
 import shutil
 import numpy as np
+from pathlib import Path
 
 import h5py
 
@@ -28,18 +29,18 @@ def make_stack_from_single_tifs(in_folder,out_folder,file_name=None,T_max=np.inf
         assert (file_name is not None), "Please provide a file_name for the resulting stack, when using a list of folders"
         fnames = []
         for folder in in_folder:
-            fnames_tmp = [os.path.join(folder,f) for f in os.listdir(folder) if f.lower().endswith('.tif') and not f.startswith('.')]
+            fnames_tmp = [Path(folder,f) for f in os.listdir(folder) if f.lower().endswith('.tif') and not f.startswith('.')]
             fnames_tmp.sort()
             fnames.extend(fnames_tmp)
         fname = file_name
     else:
-        fnames = [os.path.join(in_folder,f) for f in os.listdir(in_folder) if f.lower().endswith('.tif') and not f.startswith('.')]
+        fnames = [Path(in_folder,f) for f in os.listdir(in_folder) if f.lower().endswith('.tif') and not f.startswith('.')]
         fnames.sort()
 
         fname = file_name if file_name else os.path.splitext(fnames[0])[0][:-4]     ## assuming single recording images, where the last 4 digits indicate framenumber
 
     os.makedirs(out_folder,exist_ok=True)
-    out_path = os.path.join(out_folder,os.path.split(fname)[-1]+'.tif')
+    out_path = Path(out_folder,os.path.split(fname)[-1]+'.tif')
 
     ## get information on size of resulting stack and check for data consistency
     tif = TiffFile(fnames[0])
@@ -57,21 +58,23 @@ def make_stack_from_single_tifs(in_folder,out_folder,file_name=None,T_max=np.inf
     T = min(T_max,T_total)
 
     # if file exists already, remove first to avoid conflicts (shouldn't be necessary)
-    if os.path.exists(out_path):
+    if Path(out_path).is_file():
         os.remove(out_path)
 
     # prepare parameters of memmapped file
     d = (T,) + d
     data_type = data_type if data_type else data_type_in
-    print(f"Now merging files into a single stack {out_path} with d={d} frames ({T_total} found) and datatype {data_type}")
+    print(f"Now merging input file(s) into a tiff stack {out_path} with d={d} frames ({T_total} found) and datatype {data_type} (from {data_type_in})")
 
     if normalize:
         if str(data_type_in).startswith('float'):
             normalize = False
         else:
             max_val = np.iinfo(data_type_in).max
-    print(data_type,max_val)
-
+            # print(data_type,max_val)
+    # else:
+        # print(data_type)
+    # max_val = np.iinfo(data_type).max
     memmap_image = memmap(out_path,shape=d,dtype=data_type)
 
     # writing data to memmapped tif-file
@@ -98,10 +101,11 @@ def make_stack_from_single_tifs(in_folder,out_folder,file_name=None,T_max=np.inf
 
     if clean_after_stacking:
         print("Removing tmp files...")
-        try:
-            shutil.rmtree(in_folder)
-        except:
-            pass
+        if Path(out_path).is_file():
+            try:
+                shutil.rmtree(in_folder)
+            except:
+                pass
     return out_path
 
 
@@ -145,7 +149,7 @@ def make_stack_from_h5(in_file,out_folder,T_max=np.inf,data_type=None,normalize=
     T = min(T_max,T_total)
     
     data_type = data_type if data_type else data_type_in
-    print(f"Now writing input .h5 file into a tiff stack {out_path} with d={dims} frames ({T_total} found) and datatype {data_type}")
+    print(f"Now writing input file(s) into a tiff stack {out_path} with d={dims} frames ({T_total} found) and datatype {data_type} (from {data_type_in})")
 
     if normalize:
         if str(data_type_in).startswith('float'):
